+++ src/layouts/AuthLayout.tsx
@@ -5,7 +5,7 @@ export const AuthLayout = () => {
   return (
     <Box
       sx={{
-        minHeight: "100vh",
+        minHeight: "calc(100vh - 64px)",
         display: "flex",
         alignItems: "center",
         justifyContent: "center",